refactor users and join for backend
